Fase2/GreenMarket/backend/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1.1
djangorestframework==3.15.2
filelock==3.16.1
mysqlclient==2.2.4
pillow==10.4.0
platformdirs==4.3.6
PyMySQL==1.1.1
sqlparse==0.5.1
tzdata==2024.2
virtualenv==20.26.5
